[Feat] layout title 수정 #118
