--- src/components/energy-trade/기존코드/index.ts
@@ -1,5 +1,0 @@
-export { StatusCards } from "./StatusCards";
-export { BarChartComponent } from "./BarChart";
-export { LineChartComponent } from "./LineChart";
-export { DataTable } from "./DataTable";
-export { Forecast } from "./Forecast";
